+++ miscbotlib.h
@@ -34,5 +34,5 @@ string rgxReplace(string str1, string str2, string str3);
 // Unicode
 long fromUnicode(string str);
 string toUnicode(long val);
-short unicodeLen(char c);
+size_t unicodeLen(char c);
 vector<long> unidecode(string str);
